+++ eth-providers/src/utils/transactionReceiptHelper.ts
@@ -3,14 +3,15 @@ import { hexValue } from '@ethersproject/bytes';
 import { keccak256 } from '@ethersproject/keccak256';
 import { Formatter, TransactionReceipt } from '@ethersproject/providers';
 import type { GenericExtrinsic, i32, u64 } from '@polkadot/types';
-import type { EventRecord } from '@polkadot/types/interfaces';
+import type { EventRecord, SignedBlock } from '@polkadot/types/interfaces';
 import type { EvmLog, H160, ExitReason } from '@polkadot/types/interfaces/types';
 import { FrameSystemEventRecord } from '@polkadot/types/lookup';
 import { hexToU8a, nToU8a } from '@polkadot/util';
 import { logger } from './logger';
 import { isOrphanEvmEvent } from './utils';
 import { TransactionReceipt as TransactionReceiptSubql } from './gqlTypes';
-import { BIGNUMBER_ZERO, DUMMY_V_R_S } from '../consts';
+import { BIGNUMBER_ZERO, DUMMY_V_R_S, ORPHAN_TX_DEFAULT_INFO } from '../consts';
+import { TX } from 'src/base-provider';
 
 export interface PartialLog {
   removed: boolean;
@@ -271,3 +272,24 @@ export const subqlReceiptAdapter =
     })
     : null
   );
+
+export const receiptToTransaction = (tx: FullReceipt, block: SignedBlock): TX => {
+  const extrinsic = block.block.extrinsics.find(ex => ex.hash.toHex() === tx.transactionHash);
+
+  const extraData = extrinsic
+    ? parseExtrinsic(extrinsic)
+    : ORPHAN_TX_DEFAULT_INFO;
+
+  return {
+    blockHash: tx.blockHash,
+    blockNumber: tx.blockNumber,
+    transactionIndex: tx.transactionIndex,
+    hash: tx.transactionHash,
+    from: tx.from,
+    gasPrice: tx.effectiveGasPrice,
+    ...extraData,
+
+    // overrides `to` in parseExtrinsic, in case of non-evm extrinsic, such as dex.xxx
+    to: tx.to || null,
+  };
+};
